// angular-frontend/src/app/interfaces/vehicle-management.ts
export class VehicleManagement {
  id: number;
  vehicleNumber: string;
  driverName: string;
  vehicleType: string;
  capacity: number;
  costPerSlot: number;
  vehicleStatus: string;
  // date: string;
  slot1: string;
  slot2: string;
  slot3: string;
  companyName: string;
  bookingId:number;
  slot: string;
  retailerId: number;
  requestStatus: string;
  remainingCapacity: number;
}