// angular-frontend/src/app/interfaces/retailerDetails.ts
export class Retailerdetails{

  id: number;
  firstName: string;
  lastName: string;
  age: string;
  emergencyContact: string;
  dateOfBirth: string;
  placeOfBirth: string;
  gstIn: string;
  permanentAddress: string;
  shopAddress: string;
  mobileNo: string;
  fullName: string;
  email: string;
  phoneNo: string;
  address: string;
  docName: string;
  companyName: string;
  profilePic: string ;
  profilePicType: string;
  docPic: string;
  docPicType: string;
}